Update generator for kpoint grid generation and the molecular case
Update the protocol to set a target k-point spacing rather than an explicit
number of kpoints.

Add support for the molecular case when an input structure has the `pbc` attribute
set to [False, False, False]. In this instance, a gamma point calculatiomn with
high-cuftoff is required to obtain reasonable forces.